requirements.txt
requests==2.28.1
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-JWT==0.3.2
python-dotenv==0.20.0